* PlaylistView.m           use NSDragOperation, and some int -> NSInteger transitions         * Playlist.[h|m]           return NSUInteger for numberOfSongs instead of unsigned int           use NSUInteger with songAtIndex:           return NSUInteger for indexOfSong, and use NSUInteger for           moveSongsAtIndexes:toIndex:         * PlaylistController.m           fix format string after numberOfSongs, and minor int -> NSInteger           transitions         * Frameworks/Cynthiune/utils.h           fix format string
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2645 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Cynthiune/Playlist.h
@@ -52,11 +52,11 @@
 
 - (void) replaceSongsWithArray: (NSArray *) array;
 
-- (Song *) songAtIndex: (unsigned int) index;
-- (unsigned int) indexOfSong: (Song *) song;
+- (Song *) songAtIndex: (NSUInteger) index;
+- (NSUInteger) indexOfSong: (Song *) song;
 
-- (unsigned int) moveSongsAtIndexes: (NSArray *) indexes
-                            toIndex: (unsigned int) index;
+- (NSUInteger) moveSongsAtIndexes: (NSArray *) indexes
+                            toIndex: (NSUInteger) index;
 
 - (Song *) firstSong;
 - (Song *) lastSong;
@@ -68,7 +68,7 @@
 - (Song *) validSongAfter: (Song *) song;
 - (Song *) validSongBefore: (Song *) song;
 
-- (unsigned int) numberOfSongs;
+- (NSUInteger) numberOfSongs;
 
 - (NSNumber *) duration;
